requirements.txt
swarms
groq
python-dotenv
gradio
termcolor
pathos